feat: setup react navigation
